backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
aiomysql
alembic==1.13.1
python-dotenv==1.0.1
cryptography==41.0.5
fastapi-users[sqlalchemy]==14.0.1
pydantic==2.8.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20

# Testing
pytest==7.4.4
pytest-asyncio==0.23.8
httpx==0.26.0
pytest-cov==4.1.0
